requirements.txt
altgraph==0.17
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
PyInstaller==3.6
pylint==2.4.4
PySide2==5.14.0
python-qbittorrent==0.4.2
requests==2.22.0
shiboken2==5.14.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
